Finalized the CSS fixes/consolidation and properly commented on each segment
Consolidated a lot of elements that had similar values into one class selector that can dictate once for the batch of code that apply. Also properly commented on each element in CSS to describe purpose and target, as well as described instances where I consolidated and reorganized the CSS file.